Rename setting to LOAD_SOURCE_MAP and refactor preamble code
