+++ emcc.py
@@ -1475,7 +1475,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           shared.Settings.UBSAN_RUNTIME = 2
 
       if sanitize and '-g4' in args:
-        shared.Settings.SOURCE_MAPS = 1
+        shared.Settings.LOAD_SOURCE_MAP = 1
 
       if shared.Settings.WASM_BACKEND:
         options.js_opts = None
